chore: hoist installPreviewObserver, collectOfferIdSet, fetchToken, TABLE_SEL; flip A4-A8 backlog status (A4-A9)
