+++ ibl5/tests/e2e/helpers/csrf.ts
@@ -40,3 +40,48 @@ export async function fetchRookieOptionCsrfToken(
   }
   return match[1];
 }
+
+/**
+ * Fetch a fresh, valid `_csrf_token` from a GET-rendered form on `path`.
+ *
+ * Pass `formName` to scope extraction to a specific `<form name="...">`; the
+ * token for that form is the first `_csrf_token` after the form's opening tag.
+ * Without `formName`, the page must contain exactly one token — if it has more,
+ * extraction is ambiguous and this throws rather than silently picking one.
+ */
+export async function fetchToken(
+  request: APIRequestContext,
+  path: string,
+  formName?: string,
+): Promise<string> {
+  const resp = await request.get(path);
+  if (!resp.ok()) {
+    throw new Error(`token fetch GET ${path} failed: ${resp.status()}`);
+  }
+  let body = await resp.text();
+  if (formName === undefined) {
+    const tokenCount = (
+      body.match(/name="_csrf_token" value="[0-9a-f]+"/g) ?? []
+    ).length;
+    if (tokenCount > 1) {
+      throw new Error(
+        `ambiguous _csrf_token on ${path}: ${tokenCount} tokens found — ` +
+          `pass a formName to scope extraction to the form under test`,
+      );
+    }
+  } else {
+    const formIdx = body.indexOf(`name="${formName}"`);
+    if (formIdx === -1) {
+      throw new Error(`form name="${formName}" not found on ${path}`);
+    }
+    body = body.slice(formIdx);
+  }
+  const match = body.match(/name="_csrf_token" value="([0-9a-f]+)"/);
+  if (match === null) {
+    throw new Error(
+      `no _csrf_token found on ${path}` +
+        (formName !== undefined ? ` in form "${formName}"` : ''),
+    );
+  }
+  return match[1];
+}
